feat: review changes + moving to package + misc
