+++ pkg/output/stats/waf/regexes.json
